Cargando evidencias de trabajo ya que me lo han empezado a revisar.
I am going to publish an algorithm system used to have a multidimensional vector index. Multidimensional because it can be adapted to any number of dimensions. Vector, because... vectors, shapes, areas, etc. It allows your application to look for contained data in a decently low timespan. My target is systems with a massive numbers of elements to control.